+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Instagram like automation
 
+## 2.2
+- added location auto switcher
+- better console logs
+- show when script should be run again
+- maxExecutions per account instead of global limit
+- save latest 300 liked posts authors and skip them
+- logout via xhr request instead of clicking "logout button"
+
 ## 2.1
 - fixed logout bug